Synchronize files in React state with browser storage
