+++ package.json
@@ -26,6 +26,7 @@
     "eslint-plugin-only-warn": "1.0.3",
     "highlight.js": "11.2.0",
     "husky": "7.0.2",
+    "localforage": "1.10.0",
     "marked": "3.0.4",
     "normalize.css": "8.0.1",
     "react": "^17.0.2",
