+++ src/components/content/content.styled.ts
@@ -37,7 +37,7 @@ export const TypeArea = styled.div`
 `
 
 export const Separator = styled.div`
-  height: 100vh;
+  height: 85vh;
   width: 2px;
   background: rgba(30, 41, 59, 0.12);
 `
